Improve Makefile (.PHONY)

+++ Makefile
@@ -1,15 +1,15 @@
+.PHONY: install clean dist dist-zip serve
+
 install:
 	pip install -r requirements.txt
 
 clean:
-	rm -r site/
+	[ -d site/ ] && rm -r site/ || true
 
 dist:
 	zensical build
 
-dist-zip:
-	rm -r site/
-	zensical build
+dist-zip: clean dist
 	zip -r site/topomine-doc.zip site
 
 serve:
